% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{state.names}
\alias{state.names}
\title{list of state names and their abbreviations}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 8 rows and 2 columns.
}
\usage{
state.names
}
\description{
list of state names and their abbreviations
}
\keyword{datasets}
